Secure notify: remove public API route, use internal mailer utility
The /api/notify endpoint was publicly accessible, allowing anyone to
send arbitrary emails via the app's SMTP. Replaced it with an internal
sendNotification() utility called directly from the appointment route.
This also removes unnecessary HTTP round-trips (axios calls to own API).
